refactor(middleware): split skill and filesystem tools
